Included FastNoiseSIMD (for later replacement of libnoise) and created a primitive instanced class to start prepping for trees/rocks in scene

+++ VulkanRenderer/Water.cpp
@@ -136,8 +136,8 @@ void Water::SetupPipeline(VulkanPipeline PipelineManager, VkRenderPass renderPas
 	PipelineManager.SetFragmentShader(myPipe, loadShaderModule(device->device, "shaders/water.frag.spv"));
 	PipelineManager.SetVertexInput(myPipe, Vertex::getBindingDescription(), Vertex::getAttributeDescriptions());
 	PipelineManager.SetInputAssembly(myPipe, VK_PRIMITIVE_TOPOLOGY_TRIANGLE_LIST, 0, VK_FALSE);
-	PipelineManager.SetViewport(myPipe, viewPortWidth, viewPortHeight, 0.0f, 1.0f, 0.0f, 0.0f);
-	PipelineManager.SetScissor(myPipe, viewPortWidth, viewPortHeight, 0.0f, 0.0f);
+	PipelineManager.SetViewport(myPipe, (float)viewPortWidth, (float)viewPortHeight, 0.0f, 1.0f, 0.0f, 0.0f);
+	PipelineManager.SetScissor(myPipe, viewPortWidth, viewPortHeight, 0, 0);
 	PipelineManager.SetViewportState(myPipe, 1, 1, 0);
 	PipelineManager.SetRasterizer(myPipe, VK_POLYGON_MODE_FILL, VK_CULL_MODE_NONE, VK_FRONT_FACE_COUNTER_CLOCKWISE, VK_FALSE, VK_FALSE, 1.0f, VK_TRUE);
 	PipelineManager.SetMultisampling(myPipe, VK_SAMPLE_COUNT_1_BIT);
